langchain/requirements.txt
# LangChain Core
langchain
langchain-classic
langchain-core
langchain-community

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub
sentence_transformers

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# streamlit
streamlit

# groq
langchain-groq

pydantic 
email-validator

pypdf

beautifulsoup4

langchain-experimental

chromadb

langchain-chroma

faiss-cpu